align with rpc API
