+++ livekit-rtc/livekit/rtc/data_stream.py
@@ -41,7 +41,7 @@ class BaseStreamInfo(TypedDict):
     topic: str
     timestamp: int
     size: Optional[int]
-    attributes: Optional[Dict[str, str]]  # Optional for the extensions dictionary
+    attributes: Optional[Dict[str, str]]  # Optional for the attributes dictionary
 
 
 @dataclass
@@ -259,7 +259,7 @@ class TextStreamWriter(BaseStreamWriter):
         local_participant: LocalParticipant,
         *,
         topic: str = "",
-        extensions: Optional[Dict[str, str]] = {},
+        attributes: Optional[Dict[str, str]] = {},
         stream_id: str | None = None,
         total_size: int | None = None,
         reply_to_id: str | None = None,
@@ -268,7 +268,7 @@ class TextStreamWriter(BaseStreamWriter):
         super().__init__(
             local_participant,
             topic,
-            extensions,
+            attributes,
             stream_id,
             total_size,
             mime_type="text/plain",
@@ -313,7 +313,7 @@ class ByteStreamWriter(BaseStreamWriter):
         *,
         name: str,
         topic: str = "",
-        extensions: Optional[Dict[str, str]] = None,
+        attributes: Optional[Dict[str, str]] = None,
         stream_id: str | None = None,
         total_size: int | None = None,
         mime_type: str = "application/octet-stream",
@@ -322,7 +322,7 @@ class ByteStreamWriter(BaseStreamWriter):
         super().__init__(
             local_participant,
             topic,
-            extensions,
+            attributes,
             stream_id,
             total_size,
             mime_type=mime_type,
